delete connection added, bugs fixed, keyring removed, code cleanup
