+++ .gitignore
@@ -154,3 +154,5 @@ dmypy.json
 
 # Pyre type checker
 .pyre/
+
+connections.ini
