+++ jupysql_plugin/_widgets/__init__.py
@@ -1,0 +1,4 @@
+from jupysql_plugin._widgets._widgets import FormWidget, TableWidget
+from jupysql_plugin._widgets.connector_widget import ConnectorWidget
+
+__all__ = ["FormWidget", "TableWidget", "ConnectorWidget"]
